chore: cleanup and version bump to 4.10.0
- Renamed DisassemblerPaginatedV2 to DisassemblerPaginated
- Removed old unused components:
  - DisassemblerPaginatedNoScroll.tsx
  - MemoryViewerPaginatedV2.tsx
- Updated imports to use final component names
- Bumped version to 4.10.0 for paginated table improvements
