+++ src/components/DisassemblerPaginated.tsx
@@ -1,6 +1,8 @@
-import React, { useState, useEffect, useCallback, useRef } from 'react';
-import { WORKER_MESSAGES, MemoryRangeRequest, MemoryRangeData } from '../apple1/TSTypes';
+import React, { useState, useEffect, useCallback } from 'react';
+import { WORKER_MESSAGES, MemoryRangeRequest, MemoryRangeData, DebugData } from '../apple1/TSTypes';
 import { OPCODES } from './Disassembler';
+import PaginatedTableView from './PaginatedTableView';
+import { usePaginatedTable } from '../hooks/usePaginatedTable';
 
 interface DisassemblerProps {
     worker: Worker;
@@ -18,18 +20,37 @@ type ExecutionState = 'running' | 'paused' | 'stepping';
 const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
     const [lines, setLines] = useState<DisassemblyLine[]>([]);
     const [currentPC, setCurrentPC] = useState<number>(0);
-    const [startAddress, setStartAddress] = useState(0x0000);
-    const [addressInput, setAddressInput] = useState('0000');
     const [breakpoints, setBreakpoints] = useState<Set<number>>(new Set());
     const [executionState, setExecutionState] = useState<ExecutionState>('running');
     const [isPaused, setIsPaused] = useState(false);
     
-    const scrollContainerRef = useRef<HTMLDivElement>(null);
-    const pcRowRef = useRef<HTMLTableRowElement>(null);
+    // We fetch more than needed to ensure we have enough instructions
+    const MEMORY_CHUNK_SIZE = 512;
     
-    // Fixed size for memory fetch - enough to fill a typical view
-    const MEMORY_CHUNK_SIZE = 512; // bytes
-
+    // Use the pagination hook - note: bytesPerRow is variable for disassembly
+    const {
+        currentAddress,
+        visibleRows,
+        navigateTo,
+        containerRef,
+        contentRef,
+        getAddressRange
+    } = usePaginatedTable({
+        initialAddress: 0x0000,
+        bytesPerRow: 1, // Not really used for disassembly
+        rowHeight: 24, // Same as memory viewer
+        onDataRequest: (addr) => {
+            // Request memory for disassembly
+            const length = Math.min(MEMORY_CHUNK_SIZE, 0x10000 - addr);
+            const request: MemoryRangeRequest = { start: addr, length };
+            worker.postMessage({
+                type: WORKER_MESSAGES.GET_MEMORY_RANGE,
+                data: request
+            });
+        }
+    });
+    
+    // Disassemble memory
     const disassembleMemory = useCallback((startAddr: number, length: number, memory: number[]): DisassemblyLine[] => {
         const result: DisassemblyLine[] = [];
         let addr = startAddr;
@@ -54,7 +75,7 @@ const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
             
             // Get operand bytes
             for (let i = 1; i < opcodeInfo.bytes; i++) {
-                if (addr + i < 0x10000) {
+                if (addr + i < 0x10000 && addr + i - startAddr < memory.length) {
                     bytes.push(memory[addr + i - startAddr] || 0);
                 }
             }
@@ -62,26 +83,21 @@ const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
             // Format operand based on addressing mode
             switch (opcodeInfo.mode) {
                 case 'imp':
-                case 'acc': {
+                case 'acc':
                     operand = '';
                     break;
-                }
-                case 'imm': {
+                case 'imm':
                     operand = `#$${bytes[1]?.toString(16).padStart(2, '0').toUpperCase() || '00'}`;
                     break;
-                }
-                case 'zp': {
+                case 'zp':
                     operand = `$${bytes[1]?.toString(16).padStart(2, '0').toUpperCase() || '00'}`;
                     break;
-                }
-                case 'zpx': {
+                case 'zpx':
                     operand = `$${bytes[1]?.toString(16).padStart(2, '0').toUpperCase() || '00'},X`;
                     break;
-                }
-                case 'zpy': {
+                case 'zpy':
                     operand = `$${bytes[1]?.toString(16).padStart(2, '0').toUpperCase() || '00'},Y`;
                     break;
-                }
                 case 'abs': {
                     const absAddr = (bytes[2] || 0) << 8 | (bytes[1] || 0);
                     operand = `$${absAddr.toString(16).padStart(4, '0').toUpperCase()}`;
@@ -102,14 +118,12 @@ const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
                     operand = `($${indAddr.toString(16).padStart(4, '0').toUpperCase()})`;
                     break;
                 }
-                case 'izx': {
+                case 'izx':
                     operand = `($${bytes[1]?.toString(16).padStart(2, '0').toUpperCase() || '00'},X)`;
                     break;
-                }
-                case 'izy': {
+                case 'izy':
                     operand = `($${bytes[1]?.toString(16).padStart(2, '0').toUpperCase() || '00'}),Y`;
                     break;
-                }
                 case 'rel': {
                     const offset = bytes[1] || 0;
                     const target = addr + 2 + (offset > 127 ? offset - 256 : offset);
@@ -130,56 +144,101 @@ const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
         
         return result;
     }, []);
+    
+    // Handle worker messages
+    useEffect(() => {
+        const handleWorkerMessage = (event: MessageEvent) => {
+            if (event.data?.type === WORKER_MESSAGES.MEMORY_RANGE_DATA) {
+                const memoryData = event.data.data as MemoryRangeData;
+                const disassembly = disassembleMemory(memoryData.start, memoryData.data.length, memoryData.data);
+                // Trim to visible rows
+                const trimmedLines = disassembly.slice(0, visibleRows);
+                setLines(trimmedLines);
+            }
+            
+            if (event.data?.type === WORKER_MESSAGES.DEBUG_DATA) {
+                const debugData = event.data.data as DebugData;
+                if (debugData.cpu?.PC !== undefined) {
+                    const pc = typeof debugData.cpu.PC === 'string' 
+                        ? parseInt(debugData.cpu.PC.replace('$', ''), 16)
+                        : debugData.cpu.PC;
+                    setCurrentPC(pc);
+                }
+            }
+            
+            if (event.data?.type === WORKER_MESSAGES.BREAKPOINTS_DATA) {
+                const bpData = event.data.data as number[];
+                setBreakpoints(new Set(bpData));
+            }
+            
+            if (event.data?.type === WORKER_MESSAGES.BREAKPOINT_HIT) {
+                const hitPC = event.data.data as number;
+                navigateTo(hitPC);
+            }
+        };
 
-    // Load memory at address
-    const loadMemoryAt = useCallback((addr: number) => {
-        if (addr < 0 || addr > 0xFFFF) return;
-        
-        setStartAddress(addr);
-        setAddressInput(addr.toString(16).padStart(4, '0').toUpperCase());
-        setLines([]);
-        
-        const length = Math.min(MEMORY_CHUNK_SIZE, 0x10000 - addr);
-        const request: MemoryRangeRequest = { start: addr, length };
-        worker.postMessage({
-            type: WORKER_MESSAGES.GET_MEMORY_RANGE,
-            data: request
-        });
+        worker.addEventListener('message', handleWorkerMessage);
+        return () => worker.removeEventListener('message', handleWorkerMessage);
+    }, [worker, disassembleMemory, navigateTo, visibleRows]);
+    
+    // Listen for emulation status updates
+    useEffect(() => {
+        const handleMessage = (e: MessageEvent) => {
+            if (e.data.type === WORKER_MESSAGES.EMULATION_STATUS) {
+                const status = e.data.data;
+                if (status === 'paused') {
+                    setIsPaused(true);
+                    setExecutionState('paused');
+                } else {
+                    setIsPaused(false);
+                    setExecutionState('running');
+                }
+            }
+        };
+
+        worker.addEventListener('message', handleMessage);
+        return () => worker.removeEventListener('message', handleMessage);
     }, [worker]);
-
-    // Navigate up/down
-    const navigateUp = useCallback(() => {
-        const newAddr = Math.max(0, startAddress - MEMORY_CHUNK_SIZE);
-        loadMemoryAt(newAddr);
-        // Scroll to bottom when going up
-        setTimeout(() => {
-            if (scrollContainerRef.current) {
-                scrollContainerRef.current.scrollTop = scrollContainerRef.current.scrollHeight;
-            }
-        }, 50);
-    }, [startAddress, loadMemoryAt]);
-
-    const navigateDown = useCallback(() => {
-        const newAddr = Math.min(0xFFFF - MEMORY_CHUNK_SIZE + 1, startAddress + MEMORY_CHUNK_SIZE);
-        loadMemoryAt(newAddr);
-        // Scroll to top when going down
-        setTimeout(() => {
-            if (scrollContainerRef.current) {
-                scrollContainerRef.current.scrollTop = 0;
-            }
-        }, 50);
-    }, [startAddress, loadMemoryAt]);
-
+    
+    // Initial load and request breakpoints
+    useEffect(() => {
+        navigateTo(0);
+        worker.postMessage({ type: WORKER_MESSAGES.GET_BREAKPOINTS });
+    }, [navigateTo, worker]);
+    
+    // Request debug info periodically
+    useEffect(() => {
+        const interval = setInterval(() => {
+            worker.postMessage({ type: WORKER_MESSAGES.DEBUG_INFO, data: '' });
+        }, 100);
+        return () => clearInterval(interval);
+    }, [worker]);
+    
+    // Navigation overrides for instruction-aware navigation
+    const handleNavigateUp = useCallback(() => {
+        // Go back by approximately the same amount we're showing
+        // Use average of 2 bytes per instruction as estimate
+        const bytesToMove = Math.max(16, visibleRows * 2);
+        const newAddr = Math.max(0, currentAddress - bytesToMove);
+        navigateTo(newAddr);
+    }, [currentAddress, visibleRows, navigateTo]);
+    
+    const handleNavigateDown = useCallback(() => {
+        // Move forward by the bytes consumed by actually rendered instructions
+        if (lines.length > 0) {
+            const lastLine = lines[lines.length - 1];
+            const nextAddr = lastLine.address + lastLine.bytes.length;
+            navigateTo(Math.min(0xFFFF, nextAddr));
+        }
+    }, [lines, navigateTo]);
+    
     // Jump to PC
     const jumpToPC = useCallback(() => {
         if (currentPC !== undefined && currentPC >= 0) {
-            loadMemoryAt(currentPC);
-            setTimeout(() => {
-                pcRowRef.current?.scrollIntoView({ behavior: 'smooth', block: 'center' });
-            }, 150);
+            navigateTo(currentPC);
         }
-    }, [currentPC, loadMemoryAt]);
-
+    }, [currentPC, navigateTo]);
+    
     // Toggle breakpoint
     const toggleBreakpoint = useCallback((address: number) => {
         const newBreakpoints = new Set(breakpoints);
@@ -198,7 +257,7 @@ const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
         }
         setBreakpoints(newBreakpoints);
     }, [breakpoints, worker]);
-
+    
     // Execution controls
     const handleStep = useCallback(() => {
         setExecutionState('stepping');
@@ -223,88 +282,205 @@ const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
     const handleReset = useCallback(() => {
         worker.postMessage({ data: 'Tab', type: WORKER_MESSAGES.KEY_DOWN });
     }, [worker]);
-
-    // Handle address input
-    const handleAddressChange = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
-        setAddressInput(e.target.value.toUpperCase());
-    }, []);
-
-    const handleAddressSubmit = useCallback((e: React.KeyboardEvent<HTMLInputElement>) => {
-        if (e.key === 'Enter') {
-            const addr = parseInt(addressInput, 16);
-            if (!isNaN(addr)) {
-                loadMemoryAt(addr);
+    
+    // Calculate end address based on actual instructions
+    const endAddress = lines.length > 0 
+        ? lines[lines.length - 1].address + lines[lines.length - 1].bytes.length - 1
+        : currentAddress;
+    
+    // Get instruction info/hints
+    const getInstructionInfo = (line: DisassemblyLine): string => {
+        const opcode = line.bytes[0];
+        const opcodeInfo = OPCODES[opcode];
+        
+        if (!opcodeInfo) return '';
+        
+        let info = '';
+        
+        // Add addressing mode hint
+        const modeNames: Record<string, string> = {
+            'imp': 'Implied',
+            'acc': 'Accumulator',
+            'imm': 'Immediate',
+            'zp': 'Zero Page',
+            'zpx': 'Zero Page,X',
+            'zpy': 'Zero Page,Y',
+            'abs': 'Absolute',
+            'abx': 'Absolute,X',
+            'aby': 'Absolute,Y',
+            'ind': 'Indirect',
+            'izx': '(Indirect,X)',
+            'izy': '(Indirect),Y',
+            'rel': 'Relative'
+        };
+        
+        info = modeNames[opcodeInfo.mode] || opcodeInfo.mode;
+        
+        // Add special hints for certain instructions
+        if (line.instruction === 'BRK') {
+            info += ' • Software interrupt';
+        } else if (line.instruction === 'RTS') {
+            info += ' • Return from subroutine';
+        } else if (line.instruction === 'RTI') {
+            info += ' • Return from interrupt';
+        } else if (line.instruction === 'JSR') {
+            info += ' • Call subroutine';
+        } else if (line.instruction === 'JMP') {
+            info += ' • Jump to address';
+        } else if (line.instruction.startsWith('B') && line.instruction !== 'BIT' && line.instruction !== 'BRK') {
+            // Branch instructions
+            const branchConditions: Record<string, string> = {
+                'BCC': 'if carry clear',
+                'BCS': 'if carry set',
+                'BEQ': 'if equal (Z=1)',
+                'BMI': 'if minus (N=1)',
+                'BNE': 'if not equal (Z=0)',
+                'BPL': 'if plus (N=0)',
+                'BVC': 'if overflow clear',
+                'BVS': 'if overflow set'
+            };
+            const condition = branchConditions[line.instruction];
+            if (condition) {
+                info += ` • Branch ${condition}`;
             }
         }
-    }, [addressInput, loadMemoryAt]);
-
-    // Listen for emulation status updates
-    useEffect(() => {
-        const handleMessage = (e: MessageEvent) => {
-            if (e.data.type === WORKER_MESSAGES.EMULATION_STATUS) {
-                const status = e.data.data;
-                if (status === 'paused') {
-                    setIsPaused(true);
-                    setExecutionState('paused');
-                } else {
-                    setIsPaused(false);
-                    setExecutionState('running');
-                }
-            }
-        };
-
-        worker.addEventListener('message', handleMessage);
-        return () => worker.removeEventListener('message', handleMessage);
-    }, [worker]);
-
-    // Handle worker messages
-    useEffect(() => {
-        const handleWorkerMessage = (event: MessageEvent) => {
-            if (event.data?.type === WORKER_MESSAGES.MEMORY_RANGE_DATA) {
-                const memoryData = event.data.data as MemoryRangeData;
-                const disassembly = disassembleMemory(memoryData.start, memoryData.data.length, memoryData.data);
-                setLines(disassembly);
-            }
+        
+        return info;
+    };
+    
+    // Render functions for the table
+    const renderTableHeader = () => (
+        <thead className="bg-surface-secondary">
+            <tr className="text-text-accent">
+                <th className="text-left px-xs py-1 border-b border-border-subtle w-8">BP</th>
+                <th className="text-left px-sm py-1 border-b border-border-subtle w-20">Address</th>
+                <th className="text-left px-sm py-1 border-b border-border-subtle w-28">Bytes</th>
+                <th className="text-left px-sm py-1 border-b border-border-subtle w-48">Instruction</th>
+                <th className="text-left px-sm py-1 border-b border-border-subtle">Info</th>
+            </tr>
+        </thead>
+    );
+    
+    const renderTableRows = () => {
+        return lines.map((line, index) => {
+            const isCurrentPC = line.address === currentPC;
+            const hasBreakpoint = breakpoints.has(line.address);
+            const addressHex = line.address.toString(16).padStart(4, '0').toUpperCase();
+            const bytesHex = line.bytes.map(b => b.toString(16).padStart(2, '0').toUpperCase()).join(' ');
             
-            if (event.data?.type === WORKER_MESSAGES.DEBUG_DATA) {
-                const debugData = event.data.data;
-                if (debugData.cpu?.PC !== undefined) {
-                    setCurrentPC(debugData.cpu.PC);
-                }
-            }
+            return (
+                <tr
+                    key={`${line.address}-${index}`}
+                    className={isCurrentPC 
+                        ? 'bg-warning/20 text-warning border-l-2 border-warning' 
+                        : 'hover:bg-surface-secondary/50 transition-colors'
+                    }
+                    style={{ height: '24px' }}
+                >
+                    <td 
+                        className="px-xs py-1 align-middle cursor-pointer text-center"
+                        onClick={() => toggleBreakpoint(line.address)}
+                        title={hasBreakpoint ? "Remove breakpoint" : "Set breakpoint"}
+                    >
+                        <span className={`inline-block w-3 h-3 rounded-full ${
+                            hasBreakpoint 
+                                ? 'bg-error/80 border border-error' 
+                                : 'bg-transparent border border-border-subtle hover:border-error/50'
+                        }`}>
+                            {hasBreakpoint && <span className="text-[8px] leading-none block text-center">●</span>}
+                        </span>
+                    </td>
+                    <td className="px-sm py-1 text-data-address align-middle font-medium">
+                        {isCurrentPC && <span className="text-warning mr-1">▶</span>} 
+                        <span className="font-mono">${addressHex}</span>
+                    </td>
+                    <td className="px-sm py-1 text-data-value align-middle font-mono text-xs w-28">
+                        {bytesHex.padEnd(8, ' ')}
+                    </td>
+                    <td className="px-sm py-1 align-middle w-48">
+                        <span className="text-data-status font-medium">{line.instruction}</span>
+                        {line.operand && (
+                            <span className="text-data-value ml-2">{line.operand}</span>
+                        )}
+                    </td>
+                    <td className="px-sm py-1 align-middle text-text-secondary text-xs">
+                        {getInstructionInfo(line)}
+                    </td>
+                </tr>
+            );
+        });
+    };
+    
+    const renderExtraControls = () => (
+        <div className="flex items-center gap-xs">
+            <button
+                onClick={handleStep}
+                disabled={!isPaused}
+                className={`px-sm py-xxs text-xs font-medium rounded transition-all ${
+                    isPaused
+                        ? 'bg-data-value/10 text-data-value hover:bg-data-value/20 border border-data-value/30'
+                        : 'bg-surface-secondary text-text-disabled cursor-not-allowed border border-border-subtle'
+                }`}
+                title="Step (F10)"
+            >
+                Step
+            </button>
             
-            if (event.data?.type === WORKER_MESSAGES.BREAKPOINTS_DATA) {
-                const bpData = event.data.data as number[];
-                setBreakpoints(new Set(bpData));
-            }
+            <button
+                onClick={handleRunPause}
+                className={`px-sm py-xxs text-xs font-medium rounded transition-all ${
+                    isPaused
+                        ? 'bg-success/10 text-success hover:bg-success/20 border border-success/30'
+                        : 'bg-warning/10 text-warning hover:bg-warning/20 border border-warning/30'
+                }`}
+                title="Run/Pause (F5)"
+            >
+                {isPaused ? 'Run' : 'Pause'}
+            </button>
             
-            if (event.data?.type === WORKER_MESSAGES.BREAKPOINT_HIT) {
-                const hitPC = event.data.data as number;
-                // Jump to the breakpoint location
-                loadMemoryAt(hitPC);
-                setTimeout(() => {
-                    pcRowRef.current?.scrollIntoView({ behavior: 'smooth', block: 'center' });
-                }, 100);
-            }
-        };
-
-        worker.addEventListener('message', handleWorkerMessage);
-        return () => worker.removeEventListener('message', handleWorkerMessage);
-    }, [worker, disassembleMemory, loadMemoryAt]);
-
-    // Keyboard shortcuts
+            <button
+                onClick={handleReset}
+                className="px-sm py-xxs text-xs font-medium rounded transition-all bg-error/10 text-error hover:bg-error/20 border border-error/30"
+                title="Reset (Tab)"
+            >
+                Reset
+            </button>
+            
+            <button
+                onClick={jumpToPC}
+                className="px-sm py-xxs text-xs font-medium rounded transition-all bg-data-address/10 text-data-address hover:bg-data-address/20 border border-data-address/30"
+                title="Jump to PC (F8)"
+            >
+                →PC
+            </button>
+        </div>
+    );
+    
+    const renderStatusInfo = () => (
+        <span className={`text-xs font-medium px-xs py-xxs rounded ${
+            executionState === 'running' 
+                ? 'bg-success/20 text-success' 
+                : executionState === 'paused'
+                ? 'bg-warning/20 text-warning'
+                : 'bg-data-value/20 text-data-value'
+        }`}>
+            {executionState.toUpperCase()}
+        </span>
+    );
+    
+    // Custom address range display for disassembly
+    const getCustomAddressRange = () => {
+        if (lines.length === 0) return getAddressRange();
+        return `$${currentAddress.toString(16).padStart(4, '0').toUpperCase()}-$${endAddress.toString(16).padStart(4, '0').toUpperCase()}`;
+    };
+    
+    // Keyboard shortcuts - handled separately due to complexity
     useEffect(() => {
         const handleKeyDown = (e: KeyboardEvent) => {
-            // Navigation
-            if (e.key === 'ArrowUp') {
-                e.preventDefault();
-                navigateUp();
-            } else if (e.key === 'ArrowDown') {
-                e.preventDefault();
-                navigateDown();
-            }
-            // Execution controls
-            else if ((e.key === 'F10' || (e.key === ' ' && e.target === document.body)) && isPaused) {
+            // Don't handle shortcuts if input is focused
+            if (document.activeElement?.tagName === 'INPUT') return;
+            
+            if ((e.key === 'F10' || (e.key === ' ' && e.target === document.body)) && isPaused) {
                 e.preventDefault();
                 handleStep();
             } else if (e.key === 'F5') {
@@ -319,197 +495,26 @@ const DisassemblerPaginated: React.FC<DisassemblerProps> = ({ worker }) => {
             }
         };
         
-        const handleMessage = (e: MessageEvent) => {
-            if (e.data?.type === 'JUMP_TO_PC') {
-                jumpToPC();
-            }
-        };
-        
         window.addEventListener('keydown', handleKeyDown);
-        window.addEventListener('message', handleMessage);
-        return () => {
-            window.removeEventListener('keydown', handleKeyDown);
-            window.removeEventListener('message', handleMessage);
-        };
-    }, [currentPC, isPaused, navigateUp, navigateDown, handleStep, handleRunPause, jumpToPC, toggleBreakpoint]);
-
-    // Initial load and request breakpoints
-    useEffect(() => {
-        loadMemoryAt(0);
-        worker.postMessage({ type: WORKER_MESSAGES.GET_BREAKPOINTS });
-    }, [loadMemoryAt, worker]);
-
-    // Calculate visible range
-    const endAddress = lines.length > 0 
-        ? lines[lines.length - 1].address + lines[lines.length - 1].bytes.length - 1
-        : startAddress;
-
+        return () => window.removeEventListener('keydown', handleKeyDown);
+    }, [currentPC, isPaused, handleStep, handleRunPause, jumpToPC, toggleBreakpoint]);
+    
     return (
-        <div className="h-full flex flex-col bg-surface-primary rounded-lg border border-border-primary">
-            {/* Integrated controls header */}
-            <div className="p-sm border-b border-border-subtle flex-shrink-0">
-                <div className="flex flex-wrap items-center gap-sm">
-                    {/* Memory Navigation */}
-                    <div className="flex items-center gap-xs">
-                        <label className="text-xs text-text-secondary">Addr:</label>
-                        <input
-                            type="text"
-                            value={addressInput}
-                            onChange={handleAddressChange}
-                            onKeyDown={handleAddressSubmit}
-                            className="bg-black/40 border border-border-primary text-data-address px-xs py-xxs w-16 font-mono text-xs rounded transition-colors focus:border-border-accent focus:outline-none"
-                            placeholder="0000"
-                            maxLength={4}
-                        />
-                        <div className="flex gap-xxs">
-                            <button
-                                onClick={navigateUp}
-                                className="px-xs py-xxs text-xs bg-surface-secondary border border-border-primary rounded hover:bg-surface-hover"
-                                title="Previous page (↑)"
-                            >
-                                ↑
-                            </button>
-                            <button
-                                onClick={navigateDown}
-                                className="px-xs py-xxs text-xs bg-surface-secondary border border-border-primary rounded hover:bg-surface-hover"
-                                title="Next page (↓)"
-                            >
-                                ↓
-                            </button>
-                        </div>
-                    </div>
-
-                    <div className="h-4 w-px bg-border-subtle" />
-
-                    {/* Execution Controls */}
-                    <div className="flex items-center gap-xs">
-                        <button
-                            onClick={handleStep}
-                            disabled={!isPaused}
-                            className={`px-sm py-xxs text-xs font-medium rounded transition-all ${
-                                isPaused
-                                    ? 'bg-data-value/10 text-data-value hover:bg-data-value/20 border border-data-value/30'
-                                    : 'bg-surface-secondary text-text-disabled cursor-not-allowed border border-border-subtle'
-                            }`}
-                            title="Step (F10)"
-                        >
-                            Step
-                        </button>
-                        
-                        <button
-                            onClick={handleRunPause}
-                            className={`px-sm py-xxs text-xs font-medium rounded transition-all ${
-                                isPaused
-                                    ? 'bg-success/10 text-success hover:bg-success/20 border border-success/30'
-                                    : 'bg-warning/10 text-warning hover:bg-warning/20 border border-warning/30'
-                            }`}
-                            title="Run/Pause (F5)"
-                        >
-                            {isPaused ? 'Run' : 'Pause'}
-                        </button>
-                        
-                        <button
-                            onClick={handleReset}
-                            className="px-sm py-xxs text-xs font-medium rounded transition-all bg-error/10 text-error hover:bg-error/20 border border-error/30"
-                            title="Reset (Tab)"
-                        >
-                            Reset
-                        </button>
-                        
-                        <button
-                            onClick={jumpToPC}
-                            className="px-sm py-xxs text-xs font-medium rounded transition-all bg-data-address/10 text-data-address hover:bg-data-address/20 border border-data-address/30"
-                            title="Jump to PC (F8)"
-                        >
-                            →PC
-                        </button>
-                    </div>
-
-                    {/* Status and range */}
-                    <div className="flex items-center gap-xs ml-auto">
-                        <span className={`text-xs font-medium px-xs py-xxs rounded ${
-                            executionState === 'running' 
-                                ? 'bg-success/20 text-success' 
-                                : executionState === 'paused'
-                                ? 'bg-warning/20 text-warning'
-                                : 'bg-data-value/20 text-data-value'
-                        }`}>
-                            {executionState.toUpperCase()}
-                        </span>
-                        {lines.length > 0 && (
-                            <span className="text-xs text-text-secondary font-mono">
-                                ${startAddress.toString(16).padStart(4, '0').toUpperCase()}-${endAddress.toString(16).padStart(4, '0').toUpperCase()}
-                            </span>
-                        )}
-                    </div>
-                </div>
-                
-                {/* Keyboard shortcuts help */}
-                <div className="mt-xs text-xs text-text-muted">
-                    ↑↓: Navigate • F10: Step • F5: Run/Pause • F8: Jump to PC • F9: Breakpoint
-                </div>
-            </div>
-
-            {/* Disassembly table */}
-            <div ref={scrollContainerRef} className="flex-1 overflow-auto">
-                <table className="text-xs border-t border-border-subtle w-full bg-black/20 font-mono table-auto">
-                    <thead className="sticky top-0 z-10 bg-surface-secondary">
-                        <tr className="text-text-accent">
-                            <th className="text-left px-xs py-1 bg-surface-secondary border-b border-border-subtle w-8">BP</th>
-                            <th className="text-left px-sm py-1 bg-surface-secondary border-b border-border-subtle">Address</th>
-                            <th className="text-left px-sm py-1 bg-surface-secondary border-b border-border-subtle">Bytes</th>
-                            <th className="text-left px-sm py-1 bg-surface-secondary border-b border-border-subtle">Instruction</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {lines.map((line, index) => {
-                            const isCurrentPC = line.address === currentPC;
-                            const hasBreakpoint = breakpoints.has(line.address);
-                            const addressHex = line.address.toString(16).padStart(4, '0').toUpperCase();
-                            const bytesHex = line.bytes.map(b => b.toString(16).padStart(2, '0').toUpperCase()).join(' ');
-                            
-                            return (
-                                <tr
-                                    key={`${line.address}-${index}`}
-                                    ref={isCurrentPC ? pcRowRef : undefined}
-                                    className={isCurrentPC 
-                                        ? 'bg-warning/20 text-warning border-l-2 border-warning' 
-                                        : 'hover:bg-surface-secondary/50 transition-colors'
-                                    }
-                                >
-                                    <td 
-                                        className="px-xs py-1 align-top cursor-pointer text-center"
-                                        onClick={() => toggleBreakpoint(line.address)}
-                                        title={hasBreakpoint ? "Remove breakpoint" : "Set breakpoint"}
-                                    >
-                                        <span className={`inline-block w-3 h-3 rounded-full ${
-                                            hasBreakpoint 
-                                                ? 'bg-error/80 border border-error' 
-                                                : 'bg-transparent border border-border-subtle hover:border-error/50'
-                                        }`}>
-                                            {hasBreakpoint && <span className="text-[8px] leading-none block text-center">●</span>}
-                                        </span>
-                                    </td>
-                                    <td className="px-sm py-1 text-data-address align-top font-medium">
-                                        {isCurrentPC && <span className="text-warning mr-1">▶</span>} 
-                                        <span className="font-mono">${addressHex}</span>
-                                    </td>
-                                    <td className="px-sm py-1 text-data-value align-top font-mono text-xs">
-                                        {bytesHex}
-                                    </td>
-                                    <td className="px-sm py-1 align-top">
-                                        <span className="text-data-status font-medium">{line.instruction}</span>
-                                        {line.operand && (
-                                            <span className="text-data-value ml-1">{line.operand}</span>
-                                        )}
-                                    </td>
-                                </tr>
-                            );
-                        })}
-                    </tbody>
-                </table>
-            </div>
-        </div>
+        <PaginatedTableView
+            currentAddress={currentAddress}
+            onAddressChange={navigateTo}
+            onNavigateUp={handleNavigateUp}
+            onNavigateDown={handleNavigateDown}
+            addressRange={getCustomAddressRange()}
+            rowCount={lines.length}
+            renderTableHeader={renderTableHeader}
+            renderTableRows={renderTableRows}
+            renderExtraControls={renderExtraControls}
+            renderStatusInfo={renderStatusInfo}
+            keyboardShortcuts="↑↓: Navigate • F10: Step • F5: Run/Pause • F8: Jump to PC • F9: Breakpoint"
+            containerRef={containerRef}
+            contentRef={contentRef}
+        />
     );
 };
