+++ src/version.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = '4.9.0';
+export const APP_VERSION = '4.10.0';
